requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-admin-interface==0.28.6
django-appconf==1.0.6
django-colorfield==0.11.0
django-compressor==4.4
django-crispy-forms==2.1
django-filter==24.2
django-import-export==3.3.7
django-registration-redux==2.13
django-tables2==2.7.0
django-tinymce==4.0.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.10.0
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
